move imports to fix Netlify errors, delete old test code
