+++ src/components/CareerSwitcher.tsx
