+++ src/components/LandingPage.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import StyledButton from '../components/StyledButton'
-import CareerSwitcher from './LandingPageComponents/CareerSwitcher';
-import SimpleStepsBox from './LandingPageComponents/SimpleStepsBox';
-import ImageWithText from './LandingPageComponents/ImageWithText';
+import CareerSwitcher from './CareerSwitcher';
+import SimpleStepsBox from './SimpleStepsBox';
+import ImageWithText from './ImageWithText';
 
 // Order of applying Tailwind CSS Classes:
 // Layout Properties -> Typography -> Background/Borders -> Flexbox & Grid -> Spacing -> Utility Classes
