+++ src/components/SimpleStepsBox.tsx
